requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.11
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
pandas==2.3.2
marshmallow==4.1.0
flask-marshmallow==1.3.0
